UPDATE : 회원가입 시GRADE 테이블에 데이터 추가 구현

+++ src/main/java/deu/manito/web/entity/Grade.java
@@ -1,7 +1,7 @@
 package deu.manito.web.entity;
 
 
-import deu.manito.web.dto.Grade.GradeDto;
+import deu.manito.web.dto.grade.GradeDto;
 import lombok.*;
 
 import javax.persistence.*;
@@ -19,13 +19,12 @@ public class Grade {
     private Long id;
 
     @Column private String nickname;
-    @Column private int    score;
+    @Column private double score;
     @Column private int    helpCnt;
 
 
     public void patch(GradeDto gradeDto) {
                 this.score = gradeDto.getScore();
                 this.helpCnt = this.helpCnt+1;
-
     }
 }
